+++ changelog.md
@@ -4,7 +4,10 @@ This file contains all the notable changes done to the Ballerina GraphQL package
 ## [Unreleased]
 
 ### Fixed
+- [[#1305] Allow Enum as an Input Parameter](https://github.com/ballerina-platform/ballerina-standard-library/issues/1305)
 - [[#1250] Fix Hanging the Service when Returning Array](https://github.com/ballerina-platform/ballerina-standard-library/issues/1250)
+- [[#1307] Returning Union of Service Types from a Resource](https://github.com/ballerina-platform/ballerina-standard-library/issues/1307)
+- [[#1274] Fix Recursive Type Reference Causing Stack Overflow](https://github.com/ballerina-platform/ballerina-standard-library/issues/1274)
 
 ## [0.2.0-alpha8] - 2021-04-23